# docs/swagger.yaml
basePath: /swagger/box-manage-service
definitions:
  controllers.APIResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      timestamp:
        example: "2025-01-26T12:00:00Z"
        type: string
    type: object
...
  controllers.PaginatedResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      page:
        example: 1
        type: integer
      size:
        example: 10
        type: integer
      timestamp:
        example: "2025-01-26T12:00:00Z"
        type: string
      total:
        example: 100
        type: integer
    type: object
...
  controllers.TaskStatusResponse:
    properties:
      box_id:
        type: integer
      box_status:
        type: string
      error_message:
        type: string
      last_heartbeat:
        type: string
      max_retries:
        type: integer
      progress:
        type: number
      retry_count:
        type: integer
      start_time:
        type: string
      status:
        type: string
      stop_time:
        type: string
      task_id:
        type: string
    type: object
...
  models.RecordTask:
    properties:
      actual_duration:
        type: number
      completed_at:
        type: string
      created_at:
        description: 创建时间
        example: "2025-01-26T12:00:00Z"
        type: string
      duration:
        description: 录制时长(分钟)
        type: integer
      error_msg:
        type: string
      file_size:
        type: integer
      id:
        description: 主键ID
        example: 1
        type: integer
      name:
        type: string
      output_path:
        description: 输出信息
        type: string
      progress:
        type: integer
      started_at:
        description: 执行信息
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.RecordTaskStatus'
        description: 状态信息
      task_id:
        description: 任务信息
        type: string
      updated_at:
        description: 更新时间
        example: "2025-01-26T12:00:00Z"
        type: string
      user_id:
        type: integer
      video_source_id:
        description: 关联信息
        type: integer
    type: object
  models.RecordTaskStatus:
    enum:
    - pending
    - recording
    - completed
    - stopped
    - failed
    type: string
    x-enum-comments:
      RecordTaskStatusCompleted: 已完成
      RecordTaskStatusFailed: 失败
      RecordTaskStatusPending: 待开始
      RecordTaskStatusRecording: 录制中
      RecordTaskStatusStopped: 已停止
    x-enum-descriptions:
    - 待开始
    - 录制中
    - 已完成
    - 已停止
    - 失败
    x-enum-varnames:
    - RecordTaskStatusPending
    - RecordTaskStatusRecording
    - RecordTaskStatusCompleted
    - RecordTaskStatusStopped
    - RecordTaskStatusFailed
...
  service.CreateRecordTaskRequest:
    properties:
      duration:
        type: integer
      name:
        type: string
      user_id:
        type: integer
      video_source_id:
        type: integer
    required:
    - duration
    - name
    - user_id
    - video_source_id
    type: object
...
  service.RecordTaskStatistics:
    properties:
      completed_count:
        type: integer
      failed_count:
        type: integer
      pending_count:
        type: integer
      recording_count:
        type: integer
      stopped_count:
        type: integer
      total_count:
        type: integer
    type: object
...
info:
  contact: {}
  description: AI盒子管理系统后端服务，提供盒子管理、模型管理、任务管理、用户管理等功能
  title: AI盒子管理系统 API
  version: "1.0"
paths:
...
  /api/v1/tasks/{id}/status:
    get:
      consumes:
      - application/json
      description: 获取指定任务的当前状态信息，包括在盒子上的运行状态
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取任务状态成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/controllers.TaskStatusResponse'
              type: object
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取任务状态
      tags:
      - 任务管理
...
  /api/v1/video/record-tasks:
    get:
      consumes:
      - application/json
      description: 分页获取录制任务列表，支持状态、视频源等条件筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 任务状态
        enum:
        - pending
        - recording
        - completed
        - stopped
        - failed
        in: query
        name: status
        type: string
      - description: 视频源ID
        in: query
        name: video_source_id
        type: integer
      - description: 用户ID
        in: query
        name: user_id
        type: integer
      - description: 任务名称搜索关键词
        in: query
        name: name
        type: string
      - description: 任务ID或描述搜索关键词
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.RecordTask'
                  type: array
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取录制任务列表
      tags:
      - 录制任务管理
    post:
      consumes:
      - application/json
      description: 创建视频录制任务，仅支持实时视频流录制
      parameters:
      - description: 录制任务创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateRecordTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.RecordTask'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 创建录制任务
      tags:
      - 录制任务管理
  /api/v1/video/record-tasks/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的录制任务及其相关数据
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 删除录制任务
      tags:
      - 录制任务管理
    get:
      consumes:
      - application/json
      description: 根据ID获取录制任务的详细信息
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.RecordTask'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取录制任务详情
      tags:
      - 录制任务管理
  /api/v1/video/record-tasks/{id}/download:
    get:
      consumes:
      - application/json
      description: 下载指定录制任务的录制文件
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 录制文件
          schema:
            type: file
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 任务不存在或文件不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 下载录制文件
      tags:
      - 录制任务管理
  /api/v1/video/record-tasks/{id}/start:
    post:
      consumes:
      - application/json
      description: 启动指定的录制任务
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 启动成功
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 启动录制任务
      tags:
      - 录制任务管理
  /api/v1/video/record-tasks/{id}/stop:
    post:
      consumes:
      - application/json
      description: 停止指定的录制任务
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 停止成功
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 停止录制任务
      tags:
      - 录制任务管理
  /api/v1/video/record-tasks/statistics:
    get:
      consumes:
      - application/json
      description: 获取录制任务的统计信息，包括各状态的任务数量
      parameters:
      - description: 用户ID，不指定则统计所有用户
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/service.RecordTaskStatistics'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取录制任务统计信息
      tags:
      - 录制任务管理
...
swagger: "2.0"